+++ lib/nbf_tools/type.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
-module NbfTools::Type
-  LINK = "link"
-  FOLDER = "folder"
+module NbfTools
+  module Type
+    LINK = "link"
+    FOLDER = "folder"
+  end
 end