+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # [Changelog]
 
+## [0.3.0] - 2023-10-21
+
+- feat: implement command line tools haha `nbf_tools`
+
+## [0.2.1] - 2023-10-11
+
+- doc: quote CHANGELOG.md method name
+
 ## [0.2.0] - 2023-10-11
 
 - feat: Reomve configuration, add parse options